# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ShopLab API
  version: 0.1.0
  summary: The JSON API behind the ShopLab demo store.
  description: |
    ShopLab is a demo online store built as a test target. Everything the web UI does goes through this API.

    **Demo site. No real payments or personal data.** Data can be wiped at any time (`POST /api/test/reset`).

    ## Conventions
    - Errors always use one shape: `{"error": {"code", "message", "fieldErrors"?}}` (schema `Error`). `fieldErrors` maps a field name to a message.
    - Lists use `{"data": [], "page", "pageSize", "total"}`. A page past the end is `200` with an empty `data` and the real `total`.
    - Money is always an integer number of cents (`priceCents`, `totalCents`, ...). Only the product list price filters (`minPrice`, `maxPrice`) are in dollars.
    - Timestamps are ISO-8601 UTC strings. Dates are `YYYY-MM-DD`.
    - Status codes: 400 validation, 401 not logged in, 402 payment declined, 403 forbidden, 404 not found, 409 conflict (such as insufficient stock), 413 upload too large, 423 locked account, 429 rate limited.

    ## Authentication
    `POST /api/auth/login` sets an httpOnly cookie (`shoplab_session`) and also returns the same value as `token`. Send either the cookie or `Authorization: Bearer <token>`; if both are sent the bearer token is used.
    Endpoints under `/api/admin/*` need an admin account: `401` when not logged in, `403` for a customer.
...

...

...
  license:
    name: MIT
servers:
  - url: /
...
security: []
paths:
...
  /api/products:
    get:
      tags: [Catalogue]
      operationId: listProducts
      summary: Search and list products
      description: |
        Only active products are listed. `q` is a case-insensitive substring match on the name. Prices in the filters are
        dollars (`10` or `10.50`) and apply to the price a customer pays (the sale price when there is one).
        Unknown parameters are ignored; invalid values are a 400 with one `fieldErrors` entry per parameter.
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 100 } }
        - name: category
          in: query
          description: A category name. Repeat the parameter or comma separate to match several.
          style: form
          explode: true
          schema:
            type: array
            items: { type: string }
        - { name: subcategory, in: query, schema: { type: string } }
        - { name: minPrice, in: query, description: Dollars, schema: { type: string, example: '10.50' } }
        - { name: maxPrice, in: query, description: Dollars, schema: { type: string, example: '99' } }
        - { name: inStock, in: query, schema: { type: boolean } }
        - { name: rating, in: query, description: Minimum rounded average rating, schema: { type: integer, minimum: 1, maximum: 5 } }
        - { name: featured, in: query, schema: { type: boolean } }
        - name: sort
          in: query
          schema: { type: string, enum: [price_asc, price_desc, rating, newest] }
        - { $ref: '#/components/parameters/Page' }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 12 }
      responses:
        '200':
          description: A page of products.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductList' }
        '400': { $ref: '#/components/responses/ValidationError' }
...
  /api/geo:
    get:
      tags: [Store info]
      operationId: getGeo
      summary: Detected shipping country
      description: |
        There is no lookup. The answer is always Sweden (`SE`) unless the caller passes `?country=<code>` with one of the
        supported codes (SE, US, IN; matched ignoring case). An unknown code is a 400.
      parameters:
        - name: country
          in: query
          description: Override for the detected country.
          schema: { type: string, example: US }
      responses:
        '200':
          description: The country.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Geo' }
        '400': { $ref: '#/components/responses/ValidationError' }
...
  /api/wishlist:
    get:
      tags: [Wishlist]
      operationId: getWishlist
      summary: The user's wishlist
      description: In wishlist order. Not paged. Saved products that are no longer sold are not shown.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: The wishlist.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WishlistList' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
    post:
      tags: [Wishlist]
      operationId: addToWishlist
      summary: Add a product
      description: Idempotent. 201 when added, 200 when it was already there.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId]
              properties:
                productId: { type: integer, minimum: 1 }
      responses:
        '200':
          description: Already on the wishlist.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WishlistAdded' }
        '201':
          description: Added.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WishlistAdded' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/cart:
    get:
      tags: [Cart]
      operationId: getCart
      summary: The current cart with totals
      description: Totals use Standard shipping; the method is chosen at checkout.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: The cart.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
    delete:
      tags: [Cart]
      operationId: clearCart
      summary: Clear the cart and its coupon
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: The (now empty) cart.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Cart' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
  /api/orders:
    post:
      tags: [Checkout and orders]
      operationId: placeOrder
      summary: Place an order from the cart
      description: |
        Send either `addressId` (a saved address) or `address` (a new one). Checks run in one transaction: validation (400), stock for
        every line (409, `fieldErrors` keyed `items.<cartItemId>`), then the payment token (402). Any failure changes nothing.
        On success the order is `Processing`, stock goes down and the cart is cleared.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PlaceOrderRequest' }
      responses:
        '201':
          description: The order.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '402':
          description: The card was declined (token `tok_declined_0002`).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '409': { $ref: '#/components/responses/Conflict' }
    get:
      tags: [Checkout and orders]
      operationId: listMyOrders
      summary: The user's own orders
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/OrderStatus' }
        - name: sort
          in: query
          schema: { type: string, enum: [date_desc, date_asc, total_desc, total_asc], default: date_desc }
        - { $ref: '#/components/parameters/Page' }
        - name: pageSize
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 5 }
      responses:
        '200':
          description: A page of order summaries.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderSummaryList' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
...
  /api/addresses:
    get:
      tags: [Account]
      operationId: listAddresses
      summary: Saved addresses, default first
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      responses:
        '200':
          description: The address book. Not paged.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AddressList' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
    post:
      tags: [Account]
      operationId: createAddress
      summary: Add an address
      description: The first address becomes the default whatever `isDefault` says. At most 10 addresses (the 11th is a 409).
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AddressInput' }
      responses:
        '201':
          description: The saved address.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SavedAddress' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/contact:
    post:
      tags: [Account]
      operationId: sendContactMessage
      summary: Send a contact message
      description: Stored in the database; no email is sent. Open to everyone.
      security: [{}, { cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [topic, message, consent]
              properties:
                topic: { type: string, enum: [order, returns, product, account, feedback, other] }
                message: { type: string, minLength: 10, maxLength: 1000 }
                consent: { type: boolean, const: true }
      responses:
        '201':
          description: Stored.
          content:
            application/json:
              schema:
                type: object
                required: [id, message]
                additionalProperties: false
                properties:
                  id: { type: integer }
                  message: { type: string }
        '400': { $ref: '#/components/responses/ValidationError' }
...
  /api/admin/products:
    get:
      tags: [Admin]
      operationId: adminListProducts
      summary: List all products, active or not
      description: In ID order. `q` is a case-insensitive substring match on the name.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - { name: q, in: query, schema: { type: string, maxLength: 100 } }
        - { name: active, in: query, description: Keep only active (true) or only inactive (false) products., schema: { type: boolean } }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/AdminPageSize' }
      responses:
        '200':
          description: A page of products.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminProductList' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [Admin]
      operationId: adminCreateProduct
      summary: Create a product
      description: |
        Send JSON, or `multipart/form-data` to include an image (every value is text in a form; whole numbers and `true`/`false` are read as such,
        and an empty `salePriceCents` means no sale price). All field problems and an unusable image come back together in `fieldErrors` (400);
        an image over 2 MB is a 413. New IDs continue after the seed (61, 62, ...) and are never reused. A new product has no variants.
        Rules: name 2 to 150 characters; category one of the six categories; subcategory one of that category's (default: the first alphabetically);
        price 1 to 1 000 000 cents; sale price below the price; stock a whole number 0 to 10 000.
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AdminProductCreate' }
          multipart/form-data:
            schema: { $ref: '#/components/schemas/AdminProductForm' }
      responses:
        '201':
          description: The created product.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminProduct' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
...
  /api/admin/orders:
    get:
      tags: [Admin]
      operationId: adminListOrders
      summary: All customers' orders, newest first
      security: [{ cookieAuth: [] }, { bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          schema: { $ref: '#/components/schemas/OrderStatus' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/AdminPageSize' }
      responses:
        '200':
          description: A page of orders.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AdminOrderList' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: shoplab_session
      description: The httpOnly session cookie set by login or register.
    bearerAuth:
      type: http
      scheme: bearer
      description: The `token` returned by login or register.
    testKey:
      type: apiKey
      in: header
      name: X-Test-Key
      description: Required for /api/test/* when the server has TEST_API_KEY set.
...
  parameters:
...
    Page:
      name: page
      in: query
      description: Starts at 1. A page past the end is a 200 with an empty `data`.
      schema: { type: integer, minimum: 1, default: 1 }
    AdminPageSize:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 50, default: 10 }
  responses:
    ValidationError:
      description: Validation failed. `fieldErrors` lists every problem.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthenticated:
      description: Not logged in (or, for /api/test/*, a missing or wrong X-Test-Key).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Forbidden:
      description: Logged in but not allowed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: The request conflicts with the current state.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    PayloadTooLarge:
      description: The upload is too large (images are limited to 2 MB). `fieldErrors.image` has the message.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    ErrorCode:
      type: string
      enum:
        - VALIDATION_ERROR
        - UNAUTHENTICATED
        - INVALID_CREDENTIALS
        - PAYMENT_DECLINED
        - FORBIDDEN
        - PAYLOAD_TOO_LARGE
        - NOT_FOUND
        - CONFLICT
        - ACCOUNT_LOCKED
        - RATE_LIMITED
        - SERVICE_UNAVAILABLE
        - INTERNAL_ERROR
    Error:
      type: object
      required: [error]
      additionalProperties: false
      properties:
        error:
          type: object
          required: [code, message]
          additionalProperties: false
          properties:
            code: { $ref: '#/components/schemas/ErrorCode' }
            message: { type: string }
            fieldErrors:
              type: object
              description: Field name to message. Keys can be nested names such as `address.postalCode` or `items.<cartItemId>`.
              additionalProperties: { type: string }
    ListMeta:
      type: object
      required: [page, pageSize, total]
      properties:
        page: { type: integer, minimum: 1 }
        pageSize: { type: integer, minimum: 0 }
        total: { type: integer, minimum: 0 }
...
    Rating:
      type: object
      required: [average, count]
      additionalProperties: false
      properties:
        average: { type: [number, 'null'], description: Average rounded half up to one decimal; null with no reviews. }
        count: { type: integer }
    ProductSummary:
      allOf:
        - $ref: '#/components/schemas/ProductSummaryFields'
      unevaluatedProperties: false
    ProductSummaryFields:
      type: object
      description: The fields of a product in lists. Used on its own as ProductSummary and extended by ProductDetail and WishlistItem.
      required: [id, name, category, subcategory, priceCents, salePriceCents, stock, inStock, hasVariants, featured, imageCount, imagePath, rating, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: string }
        subcategory: { type: string }
        priceCents: { type: integer }
        salePriceCents: { type: [integer, 'null'] }
        stock: { type: integer, minimum: 0 }
        inStock: { type: boolean }
        hasVariants: { type: boolean }
        featured: { type: boolean }
        imageCount: { type: integer }
        imagePath:
          type: [string, 'null']
          description: URL path of an image uploaded by an admin; null means the generated placeholder is used.
        rating: { $ref: '#/components/schemas/Rating' }
        createdAt: { type: string, format: date-time }
...
    ProductList:
      allOf:
        - $ref: '#/components/schemas/ListMeta'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/ProductSummary' }
      unevaluatedProperties: false
...
    Geo:
      type: object
      required: [countryCode, countryName]
      additionalProperties: false
      properties:
        countryCode: { type: string, examples: [SE] }
        countryName: { type: string, examples: [Sweden] }
...
    WishlistItem:
      allOf:
        - $ref: '#/components/schemas/ProductSummaryFields'
        - type: object
          required: [position, addedAt]
          properties:
            position: { type: integer, minimum: 1 }
            addedAt: { type: string, format: date-time }
      unevaluatedProperties: false
    WishlistList:
      allOf:
        - $ref: '#/components/schemas/ListMeta'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/WishlistItem' }
      unevaluatedProperties: false
    WishlistAdded:
      type: object
      required: [productId]
      additionalProperties: false
      properties:
        productId: { type: integer }
    ShippingMethod:
      type: string
      enum: [standard, express]
...
    Cart:
      allOf:
        - $ref: '#/components/schemas/CartFields'
      unevaluatedProperties: false
    CartFields:
      type: object
      description: The cart. Used on its own as Cart and extended by Quote.
      required: [items, itemCount, coupon, totals]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/CartItem' }
        itemCount: { type: integer, minimum: 0, description: Sum of the quantities (the header badge). }
        coupon:
          oneOf:
            - $ref: '#/components/schemas/CartCoupon'
            - type: 'null'
        totals: { $ref: '#/components/schemas/CartTotals' }
...
    OrderStatus:
      type: string
      enum: [Processing, Shipped, Delivered, Cancelled]
    AddressFields:
      type: object
      properties:
        firstName: { type: string, minLength: 1, maxLength: 60 }
        lastName: { type: string, minLength: 1, maxLength: 60 }
        street: { type: string, minLength: 3, maxLength: 100 }
        city: { type: string, maxLength: 60 }
        countryCode: { type: string, enum: [SE, US, IN] }
        regionCode: { type: string, description: A region of the chosen country (see /api/countries). }
        postalCode: { type: string, description: 5 digits for SE and US, 6 for IN. }
        phone: { type: string, description: 7 to 15 digits; digits, spaces, + - and parentheses allowed. }
    AddressInput:
      allOf:
        - $ref: '#/components/schemas/AddressFields'
        - type: object
          required: [firstName, lastName, street, countryCode, regionCode, postalCode, phone]
          properties:
            label: { type: string, maxLength: 30, description: A blank label is stored as "Address". }
            isDefault: { type: boolean }
...
    SavedAddress:
      type: object
      required: [id, label, firstName, lastName, street, city, regionCode, regionName, countryCode, countryName, postalCode, phone, isDefault]
      additionalProperties: false
      properties:
        id: { type: integer }
        label: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        street: { type: string }
        city: { type: string }
        regionCode: { type: string }
        regionName: { type: [string, 'null'] }
        countryCode: { type: string }
        countryName: { type: string }
        postalCode: { type: string }
        phone: { type: string }
        isDefault: { type: boolean }
    AddressList:
      allOf:
        - $ref: '#/components/schemas/ListMeta'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/SavedAddress' }
      unevaluatedProperties: false
    PlaceOrderRequest:
      type: object
      required: [shippingMethod, deliveryDate, paymentToken, acceptTerms]
      properties:
        shippingMethod: { $ref: '#/components/schemas/ShippingMethod' }
        deliveryDate:
          type: string
          format: date
          description: From tomorrow to 14 days ahead (UTC); Standard delivery cannot be on a weekend.
        addressId: { type: integer, description: One of the user's saved addresses. Send this or `address`. }
        address:
          allOf:
            - $ref: '#/components/schemas/AddressFields'
            - type: object
              required: [firstName, lastName, street, countryCode, regionCode, postalCode, phone]
          description: A new address, stored on the order only. Send this or `addressId`.
        paymentToken:
          type: string
          enum: [tok_ok_4242, tok_declined_0002]
          description: What the payment frame posts. `tok_declined_0002` gives a 402.
        acceptTerms: { type: boolean, const: true }
    OrderItem:
      type: object
      required: [productId, variantId, name, variantLabel, unitPriceCents, quantity, lineTotalCents]
      additionalProperties: false
      properties:
        productId: { type: integer }
        variantId: { type: [integer, 'null'] }
        name: { type: string }
        variantLabel: { type: [string, 'null'] }
        unitPriceCents: { type: integer }
        quantity: { type: integer }
        lineTotalCents: { type: integer }
    OrderAddress:
      type: object
      required: [firstName, lastName, street, city, regionCode, regionName, countryCode, countryName, postalCode, phone]
      additionalProperties: false
      properties:
        firstName: { type: string }
        lastName: { type: string }
        street: { type: string }
        city: { type: string }
        regionCode: { type: string }
        regionName: { type: string }
        countryCode: { type: string }
        countryName: { type: string }
        postalCode: { type: string }
        phone: { type: string }
    Order:
      type: object
      required: [id, number, status, createdAt, deliveryDate, couponCode, shippingMethod, subtotalCents, discountCents, shippingCents, taxCents, totalCents, paymentLast4, address, items]
      additionalProperties: false
      properties:
        id: { type: integer }
        number: { type: string, pattern: '^SL-\d{8}-\d{4}$', examples: [SL-20260930-0003] }
        status: { $ref: '#/components/schemas/OrderStatus' }
        createdAt: { type: string, format: date-time }
        deliveryDate: { type: string, format: date }
        couponCode: { type: [string, 'null'] }
        shippingMethod: { $ref: '#/components/schemas/ShippingMethod' }
        subtotalCents: { type: integer }
        discountCents: { type: integer }
        shippingCents: { type: integer }
        taxCents: { type: integer }
        totalCents: { type: integer }
        paymentLast4: { type: string }
        address: { $ref: '#/components/schemas/OrderAddress' }
        items:
          type: array
          items: { $ref: '#/components/schemas/OrderItem' }
    OrderSummary:
      type: object
      required: [id, number, status, createdAt, deliveryDate, itemCount, totalCents]
      additionalProperties: false
      properties:
        id: { type: integer }
        number: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        createdAt: { type: string, format: date-time }
        deliveryDate: { type: string, format: date }
        itemCount: { type: integer }
        totalCents: { type: integer }
    OrderSummaryList:
      allOf:
        - $ref: '#/components/schemas/ListMeta'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/OrderSummary' }
      unevaluatedProperties: false
    AdminProduct:
      type: object
      required: [id, name, category, subcategory, description, priceCents, salePriceCents, stock, active, featured, hasVariants, imageCount, imagePath, createdAt]
      additionalProperties: false
      properties:
        id: { type: integer }
        name: { type: string }
        category: { type: string }
        subcategory: { type: string }
        description: { type: string }
        priceCents: { type: integer }
        salePriceCents: { type: [integer, 'null'] }
        stock: { type: integer }
        active: { type: boolean }
        featured: { type: boolean }
        hasVariants: { type: boolean }
        imageCount: { type: integer }
        imagePath: { type: [string, 'null'] }
        createdAt: { type: string, format: date-time }
    AdminProductFields:
      type: object
      properties:
        name: { type: string, minLength: 2, maxLength: 150 }
        category: { type: string, enum: [Electronics, Clothing, Home, Sports, Books, Toys], description: Matched ignoring case. }
        subcategory: { type: string, description: One of the category's subcategories (see /api/categories). Default is the first alphabetically. }
        description: { type: string, maxLength: 2000, description: A blank description is stored as "No description provided." }
        priceCents: { type: integer, minimum: 1, maximum: 1000000 }
        salePriceCents: { type: [integer, 'null'], minimum: 1, maximum: 1000000, description: Must be below `priceCents`. Null for no sale. }
        stock: { type: integer, minimum: 0, maximum: 10000 }
        active: { type: boolean, default: true }
    AdminProductCreate:
      allOf:
        - $ref: '#/components/schemas/AdminProductFields'
        - type: object
          required: [name, category, priceCents, stock]
...
    AdminProductForm:
      type: object
      description: The same fields as JSON, sent as form text, plus the image. On PATCH send only what changes.
      properties:
        name: { type: string }
        category: { type: string }
        subcategory: { type: string }
        description: { type: string }
        priceCents: { type: string, pattern: '^\d+$' }
        salePriceCents: { type: string, description: Empty for no sale price. }
        stock: { type: string, pattern: '^\d+$' }
        active: { type: string, enum: ['true', 'false'] }
        removeImage: { type: string, enum: ['true', 'false'] }
        image:
          type: string
          format: binary
          description: PNG or JPEG, at most 2 MB. Replaces the current image.
      encoding:
        image:
          contentType: image/png, image/jpeg
    AdminProductList:
      allOf:
        - $ref: '#/components/schemas/ListMeta'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/AdminProduct' }
      unevaluatedProperties: false
    AdminOrder:
      type: object
      required: [id, number, status, createdAt, deliveryDate, itemCount, totalCents, customer]
      additionalProperties: false
      properties:
        id: { type: integer }
        number: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        createdAt: { type: string, format: date-time }
        deliveryDate: { type: string, format: date }
        itemCount: { type: integer }
        totalCents: { type: integer }
        customer:
          type: object
          required: [id, name, email]
          additionalProperties: false
          properties:
            id: { type: integer }
            name: { type: string }
            email: { type: string }
    AdminOrderList:
      allOf:
        - $ref: '#/components/schemas/ListMeta'
        - type: object
          required: [data]
          properties:
            data:
              type: array
              items: { $ref: '#/components/schemas/AdminOrder' }
      unevaluatedProperties: false
...